move sql specific stuff to its own folder
